Add pp to references
Hi Will,
Just got round to reading your review. Really informative, well-written and also very witty!
To be somewhat self-serving – I don't know if you came across poppunk where we use sketches at different k-mer lengths to estimate core and accessory distances? If this is the kind of thing that makes sense in this list (as we are slightly modifying an existing technique, not coming up with a new implementation) here is a PR adding it.
Cheers,
John